refactor: improve performance when removing elements from hash maps
The original code iterated over all hash entries, collected the elements to
delete in a list and afterwards deletes all list elements in a second pass from the map.
The comment states that this is necessary due to iterator invalidation, which however
is not correct: erasing elements from QHash does not invalidate iterators.

Erasing elements from hash maps can be done more efficient in a single pass when using erase_if.
That function was added in Qt 6.1, in Qt 5 we need to roll our own.